# specs/api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: Heimwende Advisor API
  version: "1.0.0-F02"
  description: >
    Energy upgrade advisor API for the Cloover track.
    North Star: monthly_saving = current_monthly_spend − (installment + new_energy_cost).
    The four alternatives[] are cumulative ladder steps ☀️→🔋→♨️→🚗.
servers:
  - url: http://localhost:8000
    description: Local dev
paths:
...
  /api/v1/advisor/recommend:
    post:
      summary: Recommend the max-saving upgrade ladder
      description: >
        Runs the four-rung savings ladder (☀️ Solar → 🔋 Battery → ♨️ Heat pump → 🚗 EV charger)
        and returns the best scenario plus cumulative alternatives.
        Use ?fixture=<id> to return a deterministic golden payload for the demo.
      operationId: recommend
      tags:
        - advisor
      parameters:
        - in: query
          name: fixture
          required: false
          schema:
            type: string
          description: >
            If set, return the canned fixture payload with this ID (e.g. "demo-detached")
            instead of running the engine. Enables deterministic demo mode (F24).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Household"
      responses:
        "200":
          description: Savings ladder recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
              example:
                current_monthly_spend_eur: 435
                best:
                  scenario_id: "full-bundle"
                  label: "☀️ Solar + 🔋 Battery + ♨️ Heat pump + 🚗 EV charger"
                  breakdown:
                    electricity_eur_month: 124
                    heating_eur_month: 106
                    mobility_eur_month: 134
                  capex:
                    gross_eur: 42000
                    subsidy_eur: 11000
                    after_subsidy_eur: 31000
                    subsidy_note: "€22k HP − 50% KfW 458 = €11k; PV/battery 0% VAT"
                  installment_eur_month: 244
                  monthly_saving_eur: 120
                  saving_after_payoff_eur: 364
                  break_even_month: 156
                  confidence:
                    band_eur: 35
                    low_eur: 85
                    high_eur: 155
                    biggest_driver: "self-consumption ratio (autarky 0.60 ± 0.10)"
                  payback_note: "Cost-neutral early → €364/mo after loan paid off (month 156)"
                alternatives:
                  - scenario_id: "pv-only"
                    label: "☀️ Solar 9 kWp"
                    breakdown:
                      electricity_eur_month: 80
                      heating_eur_month: 0
                      mobility_eur_month: 0
                    capex:
                      gross_eur: 13050
                      subsidy_eur: 0
                      after_subsidy_eur: 13050
                      subsidy_note: "0% VAT (§12(3) UStG)"
                    installment_eur_month: 103
                    monthly_saving_eur: -24
                    saving_after_payoff_eur: 80
                    break_even_month: 180
                    confidence:
                      band_eur: 15
                      low_eur: -39
                      high_eur: -9
                      biggest_driver: "local irradiance (PVGIS ±8%)"
                    payback_note: "Mildly negative until loads electrified; feed-in alone doesn't cover loan"
                  - scenario_id: "pv-battery"
                    label: "☀️ Solar + 🔋 Battery 8 kWh"
                    breakdown:
                      electricity_eur_month: 124
                      heating_eur_month: 0
                      mobility_eur_month: 0
                    capex:
                      gross_eur: 18650
                      subsidy_eur: 0
                      after_subsidy_eur: 18650
                      subsidy_note: "0% VAT (§12(3) UStG)"
                    installment_eur_month: 147
                    monthly_saving_eur: -24
                    saving_after_payoff_eur: 124
                    break_even_month: 180
                    confidence:
                      band_eur: 20
                      low_eur: -44
                      high_eur: -4
                      biggest_driver: "dynamic tariff spread (SMARD ±€0.04/kWh)"
                    payback_note: "Battery break-even; value unlocks when HP+EV add load"
                  - scenario_id: "pv-battery-hp"
                    label: "☀️ Solar + 🔋 Battery + ♨️ Heat pump"
                    breakdown:
                      electricity_eur_month: 124
                      heating_eur_month: 106
                      mobility_eur_month: 0
                    capex:
                      gross_eur: 29650
                      subsidy_eur: 11000
                      after_subsidy_eur: 18650
                      subsidy_note: "HP: €22k − 50% KfW 458; PV+battery 0% VAT"
                    installment_eur_month: 234
                    monthly_saving_eur: -4
                    saving_after_payoff_eur: 230
                    break_even_month: 178
                    confidence:
                      band_eur: 25
                      low_eur: -29
                      high_eur: 21
                      biggest_driver: "KfW subsidy rate (modelled 50%, range 30–70%)"
                    payback_note: "Near break-even; heat pump displaces oil spend"
                  - scenario_id: "full-bundle"
                    label: "☀️ Solar + 🔋 Battery + ♨️ Heat pump + 🚗 EV charger"
                    breakdown:
                      electricity_eur_month: 124
                      heating_eur_month: 106
                      mobility_eur_month: 134
                    capex:
                      gross_eur: 42000
                      subsidy_eur: 11000
                      after_subsidy_eur: 31000
                      subsidy_note: "€22k HP − 50% KfW 458 = €11k; PV/battery 0% VAT"
                    installment_eur_month: 244
                    monthly_saving_eur: 120
                    saving_after_payoff_eur: 364
                    break_even_month: 156
                    confidence:
                      band_eur: 35
                      low_eur: 85
                      high_eur: 155
                      biggest_driver: "self-consumption ratio (autarky 0.60 ± 0.10)"
                    payback_note: "€120/mo from day one; EV is the biggest single contributor"
                upsell:
                  from_scenario_id: "pv-battery"
                  to_scenario_id: "full-bundle"
                  delta_eur_month: 144
                  reason_md: >
                    Going from PV+battery (−€24/mo) to the full bundle lands **+€120/mo**
                    because you're still burning oil + petrol that the heat pump and EV charger displace.
                    The EV charger alone adds +€124/mo by swapping petrol (€1.85/L) for cheap home charging (≈€0.20/kWh).
                explanation_md: >
                  Ihr Haus ist ideal für das volle Paket: Die Wärmepumpe ersetzt Ihren Ölheizer
                  und spart €107/mo brutto, der EV-Lader macht Ihr Auto deutlich günstiger als Benzin.
                  Solar und Batterie decken tagsüber Ihren Eigenverbrauch und stabilisieren die Kosten
                  langfristig gegen steigende Strompreise.
                proposal_copy_md: >
                  **Ihr persönlicher Energieplan — Heimwende × Cloover**

                  Mit dem vollen Paket (Solar 9 kWp · Batterie 8 kWh · Wärmepumpe · Wallbox) sparen Sie
                  **€120/mo ab Tag eins** und **€364/mo nach Kreditende**.
                  Finanzierung ab €244/mo (180 Mo., 5% p.a. nach KfW-Förderung).
                assumptions:
                  - field: "specific_yield_kwh_per_kwp"
                    value: "980 kWh/kWp"
                    source: "PVGIS fallback (PLZ 10115)"
                    editable: true
                  - field: "kfw_subsidy_rate"
                    value: "50%"
                    source: "KfW 458 base 30% + Klima-Geschwindigkeitsbonus 20%"
                    editable: true
        "501":
          description: Not implemented (F17 wires the engine)
  /api/v1/advisor/site-check:
    post:
      summary: Permit and feasibility check for an address
      description: >
        Validates roof suitability, permit obligations and energy context for the given address.
        Called before /recommend to display the green/amber feasibility panel.
      operationId: site_check
      tags:
        - advisor
      parameters:
        - in: query
          name: fixture
          required: false
          schema:
            type: string
          description: Return a canned fixture payload instead of running live checks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteCheckRequest"
      responses:
        "200":
          description: Site feasibility result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteCheckResponse"
        "501":
          description: Not implemented (F15 / F17 wires the checks)
components:
  schemas:
...
    FuelType:
      type: string
      enum:
        - OIL
        - GAS
      description: >
        Fossil heating fuel. District heating is out of scope for v1 (§3.2).
...
    Address:
      type: object
      description: Full postal address — mandatory for Site-Check roof geometry and permit lookups.
      required:
        - street
        - house_no
        - city
      properties:
        street:
          type: string
          description: Street name (Straße).
          example: "Unter den Linden"
        house_no:
          type: string
          description: House/building number (Hausnummer).
          example: "17"
        city:
          type: string
          description: City (Ort/Stadt).
          example: "Berlin"
    HeatingInput:
      type: object
      description: Current heating system details.
      required:
        - fuel
        - eur_month
      properties:
        fuel:
          $ref: "#/components/schemas/FuelType"
        eur_month:
          type: number
          description: Current average monthly heating spend in EUR.
          example: 180
    MobilityInput:
      type: object
      description: >
        Current mobility profile. km_month is the canonical quantity; eur_month is accepted
        as an alternative and converted to km by the engine (§3.3). At least one should be provided.
      required:
        - kind
      properties:
        kind:
          $ref: "#/components/schemas/CarType"
        km_month:
          type: number
          description: Average monthly distance driven in km (canonical).
          example: 1200
        eur_month:
          type: number
          description: Average monthly spend on fuel/charging in EUR (alternative to km_month).
          example: 160
    SelectionInput:
      type: object
      description: >
        À-la-carte product selection (optional stretch, §6.3). If omitted, the engine
        runs the full nested ladder and recommends the best rung.
      properties:
        pv:
          type: boolean
          default: false
        battery:
          type: boolean
          default: false
        heat_pump:
          type: boolean
          default: false
        ev:
          type: boolean
          default: false
...
    Upsell:
      type: object
      description: >
        Incremental up-sell from one scenario to the next rung up the ladder (§6.4).
        Surfaces the Δ monthly saving of upgrading from from_scenario_id to to_scenario_id.
      required:
        - from_scenario_id
        - to_scenario_id
        - delta_eur_month
        - reason_md
      properties:
        from_scenario_id:
          type: string
          description: The current (lower) scenario being compared from.
          example: "pv-battery"
        to_scenario_id:
          type: string
          description: The recommended (higher) scenario being up-sold to.
          example: "full-bundle"
        delta_eur_month:
          type: number
          description: >
            Incremental monthly saving of upgrading (to − from).
            May be large when multiple layers are skipped.
          example: 144
        reason_md:
          type: string
          description: Markdown copy explaining why the upgrade is worthwhile.
          example: >
            Going from PV+battery (−€24/mo) to the full bundle lands **+€120/mo**
            because you're still burning oil + petrol that the heat pump and EV displace.
    Assumption:
      type: object
      description: >
        A labelled assumption used in the calculation. editable:true rows are shown
        in the assumptions drawer and re-run the engine when changed (F23).
      required:
        - field
        - value
        - source
        - editable
      properties:
        field:
          type: string
          description: Field name / assumption key.
          example: "specific_yield_kwh_per_kwp"
        value:
          type: string
          description: Human-readable value string.
          example: "980 kWh/kWp"
        source:
          type: string
          description: Official source or derivation note.
          example: "PVGIS fallback (PLZ 10115)"
        editable:
          type: boolean
          description: If true, the user can override this value to tighten the confidence band.
          example: true
    ScenarioResult:
      type: object
      description: >
        One rung of the cumulative savings ladder. monthly_saving_eur is the
        CUMULATIVE net saving at this rung (installment already deducted).
        Per-layer Δ = alternatives[n].monthly_saving_eur − alternatives[n-1].monthly_saving_eur.
      required:
        - scenario_id
        - label
        - breakdown
        - capex
        - installment_eur_month
        - monthly_saving_eur
        - saving_after_payoff_eur
        - break_even_month
        - confidence
        - payback_note
      properties:
        scenario_id:
          type: string
          description: Stable identifier for this scenario (used by upsell references).
          example: "full-bundle"
        label:
          type: string
          description: Human-readable scenario name displayed in the configurator.
          example: "☀️ Solar + 🔋 Battery + ♨️ Heat pump + 🚗 EV charger"
        breakdown:
          $ref: "#/components/schemas/Breakdown"
        capex:
          $ref: "#/components/schemas/Capex"
        installment_eur_month:
          type: number
          description: Monthly loan installment in EUR (annuity on after_subsidy capex).
          example: 244
        monthly_saving_eur:
          type: number
          description: >
            Net monthly saving vs baseline = gross_saving − installment_eur_month.
            CUMULATIVE across all layers up to this rung. May be negative early rungs.
          example: 120
        saving_after_payoff_eur:
          type: number
          description: Monthly saving once the loan is fully paid off (= gross saving only).
          example: 364
        break_even_month:
          type: integer
          description: Month number at which cumulative_net first turns non-negative.
          example: 156
        confidence:
          $ref: "#/components/schemas/Confidence"
        payback_note:
          type: string
          description: Plain-language payback summary for the honest-curve display.
          example: "€120/mo from day one; EV is the biggest single contributor"
    Tier:
      type: object
      description: >
        One of three packaged offers (low / middle / high) shown side-by-side on the dashboard,
        derived from the cumulative ladder. low = cost-efficient entry; middle = best-value
        milestone; high = full bundle headlined by the long-term (lifetime_years) cumulative
        saving plus the permanent €/mo after payoff. scenario_id references a ScenarioResult in
        alternatives[]; every € figure here is grounded in that ladder (never the LLM prose).
      required:
        - id
        - name
        - tagline
        - scenario_id
        - label
        - capex_after_subsidy_eur
        - installment_eur_month
        - monthly_saving_eur
        - saving_after_payoff_eur
        - break_even_month
        - lifetime_years
        - lifetime_saving_eur
        - headline_eur
        - headline_caption
        - rationale_md
      properties:
        id:
          type: string
          enum: [low, middle, high]
          description: Tier rank.
          example: high
        name:
          type: string
          description: Marketing name for the tier card.
          example: "Future-Proof"
        tagline:
          type: string
          description: One-line pitch shown under the tier name.
          example: "Invest in the future — maximum lifetime saving."
        scenario_id:
          type: string
          description: ID of the ScenarioResult (in alternatives[]) this tier maps to.
          example: "full-bundle"
        label:
          type: string
          description: Human-readable bundle label.
          example: "☀️ Solar + 🔋 Battery + ♨️ Heat pump + 🚗 EV charger"
        capex_after_subsidy_eur:
          type: number
          description: Total capital after subsidies for this tier.
          example: 28200
        installment_eur_month:
          type: number
          description: Monthly loan installment for this tier.
          example: 244
        monthly_saving_eur:
          type: number
          description: Net monthly saving today (installment deducted). May be small/negative for low tier.
          example: 137
        saving_after_payoff_eur:
          type: number
          description: Permanent monthly saving once the loan is paid off.
          example: 392
        break_even_month:
          type: integer
          description: Month at which cumulative cash flow turns positive.
          example: 118
        lifetime_years:
          type: integer
          description: Horizon used for the long-term cumulative figure.
          example: 20
        lifetime_saving_eur:
          type: number
          description: >
            Cumulative NET saving over lifetime_years: net saving during the loan plus
            after-payoff saving for the remaining years. The high-tier headline.
          example: 64000
        headline_eur:
          type: number
          description: The single big number to feature on this tier's card.
          example: 64000
        headline_caption:
          type: string
          description: Short caption explaining what headline_eur represents.
          example: "über 20 Jahre · danach €392/Monat dauerhaft"
        rationale_md:
          type: string
          description: Markdown one-liner on why a household would pick this tier.
          example: "Das volle Paket: höchste Gesamtersparnis und maximale Unabhängigkeit."
    Recommendation:
      type: object
      description: >
        Full advisor output. best is the recommended scenario (highest monthly_saving_eur).
        alternatives[] is the FOUR-RUNG cumulative ladder in order ☀️→🔋→♨️→🚗.
        Per-layer "+€X/mo" displayed by the FE = consecutive differences of
        alternatives[].monthly_saving_eur (AC5 identity — no extra API call).
        explanation_md and proposal_copy_md are LLM prose (F16) — never the numeric authority;
        every € in them must also appear in a ScenarioResult.
      required:
        - best
        - alternatives
        - tiers
        - upsell
        - current_monthly_spend_eur
        - explanation_md
        - proposal_copy_md
        - assumptions
      properties:
        best:
          $ref: "#/components/schemas/ScenarioResult"
        alternatives:
          type: array
          description: The four cumulative ladder rungs in order (☀️→🔋→♨️→🚗).
          items:
            $ref: "#/components/schemas/ScenarioResult"
          minItems: 1
        tiers:
          type: array
          description: >
            Three packaged offers (low / middle / high) derived from the ladder, shown
            side-by-side on the dashboard. Always ordered low → middle → high.
          items:
            $ref: "#/components/schemas/Tier"
          minItems: 1
        upsell:
          $ref: "#/components/schemas/Upsell"
        current_monthly_spend_eur:
          type: number
          description: >
            Baseline monthly spend before any upgrades (electricity + heating + mobility).
            Drives the before/after display (F21).
          example: 435
        explanation_md:
          type: string
          description: >
            LLM-generated Markdown paragraph explaining why this config fits this home (F16).
            Plain German for the demo. NEVER the numeric source of truth.
          example: "Ihr Haus ist ideal für das volle Paket..."
        proposal_copy_md:
          type: string
          description: >
            LLM-generated Markdown proposal copy for the CTA panel (F16).
            NEVER the numeric source of truth.
          example: "**Ihr persönlicher Energieplan...**"
        assumptions:
          type: array
          description: Labelled assumptions used in the calculation (F23 drawer).
          items:
            $ref: "#/components/schemas/Assumption"

    # ── Site-Check ─────────────────────────────────────────────────────────────
    SiteCheckRequest:
      type: object
      description: >
        Subset of Household used for the site-check call (§14.2).
        Accepts the address plus floor_area_m2 and building_year for roof sizing and heat-load.
      required:
        - address
        - plz
        - floor_area_m2
        - building_year
      properties:
        address:
          $ref: "#/components/schemas/Address"
        plz:
          type: string
          description: 5-digit German postcode.
          example: "10115"
        floor_area_m2:
          type: integer
          description: Living area in square metres.
          example: 140
        building_year:
          type: integer
          description: Year of construction.
          example: 1985
    FeasibilityFlag:
      type: object
      description: >
        One permit / obligation check result (§4). product names the product category,
        check names the specific rule, status is the traffic light, message is human-readable.
      required:
        - product
        - check
        - status
        - message
      properties:
        product:
          type: string
          description: Product category (e.g. "Solar PV", "Heat pump", "EV charger", "Battery").
          example: "Solar PV"
        check:
          type: string
          description: Name of the specific rule or check performed.
          example: "Building permit (Baugenehmigung)"
        status:
          $ref: "#/components/schemas/FeasibilityStatus"
        message:
          type: string
          description: Human-readable result message.
          example: "Roof PV is verfahrensfrei — no permit needed (LBO)"
    EnergyContext:
      type: object
      description: Location-resolved energy parameters returned by the site-check (F15).
      required:
        - lat
        - lon
        - specific_yield_kwh_per_kwp
        - retail_price_eur_kwh
        - grid_fee_eur_kwh
        - climate_zone
        - mastr_neighbour_count
      properties:
        lat:
          type: number
          description: Latitude resolved from the address.
          example: 52.5163
        lon:
          type: number
          description: Longitude resolved from the address.
          example: 13.3777
        specific_yield_kwh_per_kwp:
          type: number
          description: Annual PV yield per kWp for this location (kWh/kWp/yr).
          example: 980
        retail_price_eur_kwh:
          type: number
          description: Local retail electricity price in EUR/kWh.
          example: 0.37
        grid_fee_eur_kwh:
          type: number
          description: Local grid fee in EUR/kWh (per-PLZ overlay from Netztransparenz).
          example: 0.07
        climate_zone:
          type: string
          description: DE climate zone identifier (used for heat-load table, §5.3).
          example: "DE-4"
        mastr_neighbour_count:
          type:
            - integer
            - "null"
          description: >
            Number of installed PV systems in this PLZ from MaStR (social proof).
            null if unknown (no MaStR data for this PLZ).
          example: 47
    SiteCheckResponse:
      type: object
      description: Permit and feasibility result for the site-check pre-step.
      required:
        - roof_ok
        - feasibility_flags
        - energy_context
        - assumptions
      properties:
        roof_ok:
          type: boolean
          description: >
            True if the roof is viable for PV (no Denkmalschutz block, no structural flag).
            Amber = viable with caveats (user should confirm heritage listing).
          example: true
        feasibility_flags:
          type: array
          description: Per-product permit and obligation flags (§4 table).
          items:
            $ref: "#/components/schemas/FeasibilityFlag"
        energy_context:
          $ref: "#/components/schemas/EnergyContext"
        assumptions:
          type: array
          description: Labelled assumptions used in the site-check (editable → re-run).
          items:
            $ref: "#/components/schemas/Assumption"

    # ── Request: Household (the full intake schema — the seam) ─────────────────
    Household:
      type: object
      description: >
        Customer intake profile. This is the primary API seam (F02): every downstream
        feature (F05–F11, F17, F18) builds against this exact field set.
        address + plz + floor_area_m2 + building_year + occupants + electricity_eur_month
        + heating + mobility are ALL required.
      required:
        - address
        - plz
        - floor_area_m2
        - building_year
        - occupants
        - electricity_eur_month
        - heating
        - mobility
      properties:
        address:
          $ref: "#/components/schemas/Address"
        plz:
          type: string
          description: >
            5-digit German postcode. Carried alongside address because PLZ independently
            drives irradiance, grid fees, climate zone, and prices (§3.1, labelled assumption).
          example: "10115"
        floor_area_m2:
          type: integer
          description: Living floor area in square metres. Required for heat-load calculation (L3).
          example: 140
        building_year:
          type: integer
          description: Year of construction. Drives heat-load factor (§10 table) for L3.
          example: 1985
        occupants:
          type: integer
          description: Number of occupants. Drives load profile and consumption scaling.
          example: 3
        electricity_eur_month:
          type: number
          description: Average monthly electricity spend in EUR before any upgrades.
          example: 95
        heating:
          $ref: "#/components/schemas/HeatingInput"
        mobility:
          $ref: "#/components/schemas/MobilityInput"
        existing_pv_kwp:
          type: number
          default: 0
          description: >
            Installed PV capacity in kWp. 0 = no existing PV.
            If > 0, the ladder credits only the incremental yield above this.
          example: 0
        existing_battery_kwh:
          type: number
          default: 0
          description: >
            Installed battery capacity in kWh. 0 = no existing battery.
            If > 0, the ladder adds only the delta up to the recommended size.
          example: 0
        existing_heatpump_year:
          type:
            - integer
            - "null"
          description: >
            Year the existing heat pump was installed, or null if no HP present.
            null → fossil/no-HP case (Layer 3 = Case A fossil → new HP).
            If set: age ≥ 12 yrs or est. SCOP < 3.0 → Layer 3 = Case B efficiency upgrade;
            modern HP → Layer 3 Δ = 0 (not offered).
          default: null
          example: null
        existing_heatpump_power_kw:
          type:
            - number
            - "null"
          exclusiveMinimum: 0
          description: >
            Optional rated thermal output of the existing heat pump in kW.
            Overrides the area-method estimate for the Case-B baseline.
          default: null
          example: 8.0
        existing_heatpump_scop:
          type:
            - number
            - "null"
          exclusiveMinimum: 0
          description: >
            Optional measured or nameplate SCOP of the existing heat pump.
            Overrides the age-regression fallback for the Case-B baseline.
          default: null
          example: 2.5
        existing_ev:
          type: boolean
          default: false
          description: >
            True if the household already drives an EV. Changes the mobility baseline
            from fuel cost to public charging cost (§3.2).
          example: false
        existing_ev_charger:
          type: boolean
          default: false
          description: >
            True if the household already has a home wallbox.
            EV=true + charger=false → Layer 4 = Case B wallbox-only offer.
            EV=true + charger=true → Layer 4 Δ = 0 (not offered).
          example: false
        selection:
          $ref: "#/components/schemas/SelectionInput"
          description: >
            À-la-carte selection (optional). If omitted, the engine runs the full
            nested ladder and picks the best rung (recommended mode).